+++ src/ESPJsonDBCompressor.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <ESPJsonDB.h>
+
+#if !__has_include(<ESPCompressor.h>)
+#error "ESPJsonDBCompressor.h requires ESPCompressor to be available in the include path."
+#endif
+
+#include <ESPCompressor.h>